Remove an unused function and move fork to Intra
