+++ services/brig/src/Brig/App.hs
@@ -64,8 +64,6 @@ module Brig.App
     -- * App Monad
     AppT,
     runAppT,
-    runAppResourceT,
-    fork,
     locationOf,
     viewFederationDomain,
     qualifyLocal,
@@ -635,30 +633,6 @@ runAppT e (AppT ma) =
     . nowToIOAction (_currentTime e)
     . passwordResetStoreToCodeStore
     $ runReaderT ma e
-
-runAppResourceT ::
-  ResourceT (AppT BrigCanonicalEffects) a ->
-  (AppT BrigCanonicalEffects) a
-runAppResourceT ma = do
-  e <- ask
-  liftIO . runResourceT $ transResourceT (runAppT e) ma
-
-fork ::
-  (MonadIO m, MonadUnliftIO m, MonadReader Env m) =>
-  Maybe UserId ->
-  m a ->
-  m ()
-fork u ma = do
-  g <- view applog
-  r <- view requestId
-  let logErr e = Log.err g $ request r ~~ user u ~~ msg (show e)
-  withRunInIO $ \lower ->
-    void . liftIO . forkIO $
-      either logErr (const $ return ())
-        =<< runExceptT (syncIO $ lower ma)
-  where
-    request = field "request" . unRequestId
-    user = maybe id (field "user" . toByteString)
 
 locationOf :: (MonadIO m, MonadReader Env m) => IP -> m (Maybe Location)
 locationOf ip =
